#13659 #13661 Release 0.33
git-svn-id: https://dbentry.svn.codeplex.com/svn@28202 81e6e737-5a20-4d0f-a97c-cb01297b46e0

+++ src/WebDebugLib/Models/Book.cs
@@ -17,7 +17,7 @@ namespace DebugLib.Models
     {
         [Length(30)] public abstract string Name { get; set; }
         public abstract DateTime BuyDate { get; set; }
-        public abstract decimal Price { get; set; }
+        public abstract float Price { get; set; }
         public abstract BookCatagory Catagory { get; set; }
         public abstract bool Read { get; set; }
         [SpecialName] public abstract DateTime CreatedOn { get; set; }
